+++ packages/remote-config/ios/RNFBConfig/RNFBConfigModule.m
@@ -139,6 +139,8 @@ RCT_EXPORT_METHOD(setConfigSettings:
     : (RCTPromiseResolveBlock) resolve
     : (RCTPromiseRejectBlock) reject
 ) {
+  FIRRemoteConfigSettings *remoteConfigSettings =
+      [[FIRRemoteConfigSettings alloc] init];
 
   if ([configSettings objectForKey:@"minimumFetchInterval"]) {
     remoteConfigSettings.minimumFetchInterval = [configSettings[@"minimumFetchInterval"] doubleValue];
@@ -159,19 +161,6 @@ RCT_EXPORT_METHOD(setDefaults:
 ) {
   [[FIRRemoteConfig remoteConfig] setDefaults:defaults];
   resolve([self resultWithConstants:[NSNull null]]);
-}
-
-RCT_EXPORT_METHOD(setDefaultsFromResource:
-  (NSString *) fileName
-    : (RCTPromiseResolveBlock) resolve
-    : (RCTPromiseRejectBlock) reject) {
-  if ([[NSBundle mainBundle] pathForResource:fileName ofType:@"plist"] != nil) {
-    [[FIRRemoteConfig remoteConfig] setDefaultsFromPlistFileName:fileName];
-    resolve([self resultWithConstants:[NSNull null]]);
-  } else {
-    [RNFBSharedUtils rejectPromiseWithUserInfo:reject userInfo:[@{@"code": @"resource_not_found",
-        @"message": @"The specified resource name was not found."} mutableCopy]];
-  }
 }
 
 #pragma mark -