+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## remote-config 
  
-- removed `isDeveloperModeEnabled` from `remoteConfig().settings = { isDeveloperModeEnabled:true }` for both iOS & Android;
+- removed `isDeveloperModeEnabled` from `remoteConfig().settings = { isDeveloperModeEnabled:true }` for both iOS & Android. Not in web spec.
+- removed `remoteConfig().setDefaults({})` & replaced with `remoteConfig().defaultConfig = {}` which is a web property.
+- removed `remoteConfig().setDefaultsFromResource({})`, not in the web spec.
 
 # [6.1.0](https://github.com/invertase/react-native-firebase/compare/v6.0.4...v6.1.0) (2019-11-26)